+++ picky-asn1-x509/src/attribute.rs
@@ -1,7 +1,8 @@
 #[cfg(feature = "pkcs7")]
 use crate::pkcs7::content_info::SpcSpOpusInfo;
 use crate::{oids, Extension, Extensions};
-use picky_asn1::wrapper::{Asn1SequenceOf, Asn1SetOf, ObjectIdentifierAsn1, OctetStringAsn1};
+use picky_asn1::date::UTCTime;
+use picky_asn1::wrapper::{Asn1SequenceOf, Asn1SetOf, ObjectIdentifierAsn1, OctetStringAsn1, UTCTimeAsn1};
 use serde::{de, ser};
 
 pub type Attributes = Asn1SequenceOf<Attribute>;
@@ -13,7 +14,7 @@ pub type Attributes = Asn1SequenceOf<Attribute>;
 ///
 /// `spcSpOpusInfo` is behind the `pkcs7` feature.
 ///
-/// `contentType`, `messageDigest` and `spcSpOpusInfo` are used for [microsoft authenticode]
+/// `contentType`, `messageDigest`, `spcSpOpusInfo` and `SigningTime` are used for [microsoft authenticode]
 /// (http://download.microsoft.com/download/9/c/5/9c5b2167-8017-4bae-9fde-d599bac8184a/Authenticode_PE.docx)
 #[derive(Clone, Debug, PartialEq)]
 pub enum AttributeValues {
@@ -24,6 +25,7 @@ pub enum AttributeValues {
     ContentType(Asn1SetOf<ObjectIdentifierAsn1>),
     SpcStatementType(Asn1SetOf<Asn1SequenceOf<ObjectIdentifierAsn1>>),
     MessageDigest(Asn1SetOf<OctetStringAsn1>),
+    SigningTime(Asn1SetOf<UTCTimeAsn1>),
     #[cfg(feature = "pkcs7")]
     SpcSpOpusInfo(Asn1SetOf<SpcSpOpusInfo>),
     Custom(picky_asn1_der::Asn1RawDer), // fallback
@@ -42,6 +44,27 @@ impl Attribute {
             value: AttributeValues::Extensions(Asn1SetOf(vec![Extensions(extensions)])),
         }
     }
+
+    pub fn new_content_type_pkcs7() -> Self {
+        Self {
+            ty: oids::content_type().into(),
+            value: AttributeValues::ContentType(vec![oids::pkcs7().into()].into()),
+        }
+    }
+
+    pub fn new_signing_time(signing_time: UTCTime) -> Self {
+        Self {
+            ty: oids::signing_time().into(),
+            value: AttributeValues::SigningTime(vec![signing_time.into()].into()),
+        }
+    }
+
+    pub fn new_message_digest(digest: Vec<u8>) -> Self {
+        Self {
+            ty: oids::message_digest().into(),
+            value: AttributeValues::MessageDigest(vec![digest.into()].into()),
+        }
+    }
 }
 
 impl ser::Serialize for Attribute {
@@ -57,6 +80,7 @@ impl ser::Serialize for Attribute {
             AttributeValues::Custom(der) => seq.serialize_element(der)?,
             AttributeValues::ContentType(oid) => seq.serialize_element(oid)?,
             AttributeValues::MessageDigest(octet_string) => seq.serialize_element(octet_string)?,
+            AttributeValues::SigningTime(signing_time) => seq.serialize_element(signing_time)?,
             #[cfg(feature = "pkcs7")]
             AttributeValues::SpcSpOpusInfo(spc_sp_opus_info) => seq.serialize_element(spc_sp_opus_info)?,
             AttributeValues::SpcStatementType(spc_statement_type) => seq.serialize_element(spc_statement_type)?,
@@ -97,6 +121,7 @@ impl<'de> de::Deserialize<'de> for Attribute {
                     oids::MESSAGE_DIGEST => {
                         AttributeValues::MessageDigest(seq_next_element!(seq, Attribute, "an octet string"))
                     }
+                    oids::SIGNING_TIME => AttributeValues::SigningTime(seq_next_element!(seq, Attribute, "UTCTime")),
                     #[cfg(feature = "pkcs7")]
                     oids::SPC_SP_OPUS_INFO_OBJID => {
                         AttributeValues::SpcSpOpusInfo(seq_next_element!(seq, Attribute, "an SpcSpOpusInfo object"))