+++ picky-asn1-x509/CHANGELOG.md
@@ -7,16 +7,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-### Added 
+### Added
 
 - Support for Authenticode timestamp deserialization/serialization
 - CTL implementation behind `ctl` feature
 - New `SpcSipInfo` struct
+- Add serialization/deserialization of Authenticode `TimestampRequest`
+- Add timestamp request OID
+- Add a few methods for creating an Attribute without usage low-level API:
+  - `Attribute::new_content_type_pkcs7`
+  - `Attribute::new_signing_time`
+  - `Attribute::new_message_digest`
+- Add `EncapsulatedContentInfo::new_pkcs7_data` method
 
 ### Changed
 
 - (Breaking) `ShaVariant` enum is extended for MD5 and SH1 algorithms
 - (Breaking) Add `SpcStatementType` variant in `AttributeValues` enum
+- (Breaking) Add `SigningTime` variant in `AttributeValues` enum
 - `SpcAttributeAndOptionalValue` now supports both `SpcPeImageData` and `SpcSipInfo` values
 
 ### Fixed